requirements.txt
streamlit
opencv-python
numpy
Pillow
streamlit-image-coordinates
